requirements.txt
streamlit
streamlit-lottie
google-cloud-firestore
streamlit-option-menu
requests